pyproject.toml
[project]
name = "ranoparse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.1",
    "pillow>=11.3.0",
    "pypub3>=2.0.9",
    "tqdm>=4.67.1",
]
